+++ docs/src/references/changelog.rst
@@ -31,6 +31,7 @@ Fixed
 * Fix inconsistent ``cutoff`` in neighbor list example
 =======
 * Fixed consistency of dtype and device in the SplinePotential class
+* Fix inconsistent ``cutoff`` in neighbor list example
 * All calculators now check if the cell is zero if the potential is range-separated